update lib button

+++ library/src/main/java/com/gemminiii/library/Button/DefaultKvantMaterialButton/core/ButtonText.kt
@@ -10,7 +10,7 @@ interface ButtonText {
         button: MaterialButton,
         text: String?,
         textColor: ColorStateList?,
-        textSize: Float,
+        textSize: Int,
         typeface: Typeface?
     )
     fun updateTextColor(button: Button, colorStateList: ColorStateList)
